+++ src/Linting/ErrorList/CssErrorListListener.cs
@@ -36,6 +36,12 @@ public class CssErrorListListener : ITextViewCreationListener
                 view.Closed += ViewClosed;
 
                 var file = view.TextBuffer.GetFileName();
+
+                if (file is null)
+                {
+                    return;
+                }
+                
                 var tableDataSource = new TableDataSource(Vsix.Name + file);
 
                 var project = ThreadHelper.JoinableTaskFactory.Run(() => PhysicalFile.FromFileAsync(file)).ContainingProject;